% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapextent}
\alias{mapextent}
\title{Map extent}
\format{
An sf bbox object with four named numerics.
}
\usage{
mapextent
}
\description{
A boundary box for the project area.
}
\keyword{datasets}
