% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maplayout.relief}
\alias{maplayout.relief}
\title{Relief background map}
\format{
ggplot parameters
}
\usage{
maplayout.relief
}
\description{
A set of ggplot parameter for a DEM background map
}
\keyword{datasets}
