% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maptheme}
\alias{maptheme}
\title{ggplot map theme}
\format{
ggplot theme parameters.
}
\usage{
maptheme
}
\description{
A set of ggplot theme parameters to set a map background theme.
}
\keyword{datasets}
