% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relief}
\alias{relief}
\title{Relief elevation data}
\format{
A data frame with 477144 rows and 3 variables:
\describe{
\item{value}{elevation in metres ASL}
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\source{
\url{http://www.naturalearthdata.com}
}
\usage{
relief
}
\description{
A dataset the elevation of point in XYZ coordinates.
}
\keyword{datasets}
