% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate_plots.R
\name{ridgeplot}
\alias{ridgeplot}
\title{Create a ridge plot}
\usage{
ridgeplot(
  data,
  x.var,
  y.var,
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  ...
)
}
\arguments{
\item{data}{dataset}

\item{x.var}{a numeric variable}

\item{y.var}{a factor variable}

\item{title}{plot title}

\item{subtitle}{subtitle}

\item{xlabel}{x axis label}

\item{ylabel}{y axis label}
}
\value{
A ridge plot of the two variables
}
\description{
This function creates a ridge plot using the packages ggplot2 and ggridges.
The ridge plot allows you to plot the distribution of a numeric variable
separated by the levels of a factor.
}
