% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Rivers}
\format{
A data frame with 6 rows and 10 variables:
\describe{
\item{scalerank}{scale rank}
\item{featurecla}{feature class}
\item{name}{name}
\item{min_zoom}{minimum zoom}
\item{min_label}{min label}
\item{label}{label}
\item{wikidataid}{wikidataid}
\item{wdid_score}{wdid score}
\item{ne_id}{ne_id}
\item{geometry}{geometry}
}
}
\source{
\url{http://www.naturalearthdata.com}
}
\usage{
rivers
}
\description{
SF object containing all major river lines.
}
\keyword{datasets}
