% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate_plots.R
\name{scatter}
\alias{scatter}
\title{Create a scatter plot with a regression line}
\usage{
scatter(
  data,
  x.var,
  y.var,
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  ...
)
}
\arguments{
\item{data}{dataset}

\item{x.var}{a numeric variable}

\item{y.var}{a numeric variable}

\item{title}{plot title}

\item{subtitle}{subtitle}

\item{xlabel}{x axis label}

\item{ylabel}{y axis label}
}
\value{
A scatter plot of the two variables
}
\description{
This function creates a scatter plot using the package ggplot2.
The scatter plot allows you to plot the distribution of two numeric variables
and adds a regression line to indicate their correlation.
}
